add waiting approval status and rename document template files to be more readable
